refactor: StackdriverUtils to LoggerUtils
