+++ config/application.rb
@@ -3,7 +3,7 @@
 require_relative 'boot'
 
 require 'rails/all'
-require_relative '../lib/stackdriver_utils'
+require 'logger_utils'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
